+++ src/topology/mod.rs
@@ -2,6 +2,6 @@ mod edge;
 mod topology;
 mod shape;
 
-pub use edge::DirectedEdge;
+pub use edge::{DirectedEdge, EdgeKind, EdgeExtensions};
 pub use shape::*;
-pub use topology::{Topology, TopologyMetrics};
+pub use topology::{Topology, TopologyMetrics, ValidationLevel};